removed appwrite api calls
